+++ config/webpack/loaders/babel.js
@@ -4,7 +4,8 @@ module.exports = {
   use: {
     loader: 'babel-loader',
       options: {
-      presets: ['env', 'react']
+      presets: ['env', 'react'],
+      plugins: ['transform-object-rest-spread']
     }
   }
 };