+++ hbw/app/models/hbw/fields/user.rb
@@ -2,6 +2,12 @@ module HBW
   module Fields
     class User < Ordinary
       self.default_data_type = :string
+
+      def as_json
+        json = super.dup
+        json[:variables] = task.definition['variables']
+        json
+      end
     end
   end
 end