+++ spec/hbw/features/select_spec.rb
@@ -30,9 +30,9 @@ feature 'Check select with', js: true do
 
     click_and_wait 'ORD-3'
 
-    expect(page).to have_content  'ORD-3'
-    expect(page).to have_selector "[name='homsOrderDataSelect']"
-    expect(select2_text('homsOrderDataSelect')).to eq placeholder
+    expect(page).to have_content 'ORD-3'
+    expect_r_select_presence('homsOrderDataSelect')
+    expect(r_select_placeholder('homsOrderDataSelect')).to eq placeholder
     expect_widget_presence
   end
 
@@ -43,9 +43,9 @@ feature 'Check select with', js: true do
 
     click_and_wait 'ORD-4'
 
-    expect(page).to have_content  'ORD-4'
-    expect(page).to have_selector "[name='homsOrderDataSelect']"
-    expect(select2_text('homsOrderDataSelect')).to eq first_value
+    expect(page).to have_content 'ORD-4'
+    expect_r_select_presence('homsOrderDataSelect')
+    expect(r_select_single_value('homsOrderDataSelect')[:label]).to eq first_value
     expect_widget_presence
   end
 
@@ -56,9 +56,9 @@ feature 'Check select with', js: true do
 
     click_and_wait 'ORD-5'
 
-    expect(page).to have_content  'ORD-5'
-    expect(page).to have_selector "[name='homsOrderDataSelect']"
-    expect(select2_text('homsOrderDataSelect')).to eq 'Not selected'
+    expect(page).to have_content 'ORD-5'
+    expect_r_select_presence('homsOrderDataSelect')
+    expect(r_select_placeholder('homsOrderDataSelect')).to eq ''
     expect_widget_presence
   end
 
@@ -69,9 +69,9 @@ feature 'Check select with', js: true do
 
     click_and_wait 'ORD-6'
 
-    expect(page).to have_content  'ORD-6'
-    expect(page).to have_selector "[name='homsOrderDataSelect']"
-    expect(select2_text('homsOrderDataSelect')).to eq first_value
+    expect(page).to have_content 'ORD-6'
+    expect_r_select_presence('homsOrderDataSelect')
+    expect(r_select_single_value('homsOrderDataSelect')[:label]).to eq first_value
     expect_widget_presence
   end
 
@@ -82,11 +82,13 @@ feature 'Check select with', js: true do
 
     click_and_wait 'ORD-7'
 
-    expect(page).to have_content  'ORD-7'
-    expect(page).to have_selector "[name='homsOrderDataSelect']"
-    expect(select2_text('homsOrderDataSelect')).to   eq first_value
-    expect(select_options('homsOrderDataSelect')).to eq ['Option 1', 'Option 2']
-    expect(select_values('homsOrderDataSelect')).to  eq ['112233', '445566']
+    expect(page).to have_content 'ORD-7'
+    expect_r_select_presence('homsOrderDataSelect')
+    expect(r_select_single_value('homsOrderDataSelect')).to eq({label: 'Option 1', value: '112233'})
+
+    set_r_select_option('homsOrderDataSelect', 'Option 2')
+
+    expect(r_select_single_value('homsOrderDataSelect')).to eq({label: 'Option 2', value: '445566'})
     expect_widget_presence
   end
 
@@ -97,11 +99,13 @@ feature 'Check select with', js: true do
 
     click_and_wait 'ORD-8'
 
-    expect(page).to have_content  'ORD-8'
-    expect(page).to have_selector "[name='homsOrderDataSelect']"
-    expect(select2_text('homsOrderDataSelect')).to   eq first_value
-    expect(select_options('homsOrderDataSelect')).to eq ['Option 1', 'Option 2']
-    expect(select_values('homsOrderDataSelect')).to  eq ['Option 1', 'Option 2']
+    expect(page).to have_content 'ORD-8'
+    expect_r_select_presence('homsOrderDataSelect')
+    expect(r_select_single_value('homsOrderDataSelect')).to eq({label: 'Option 1', value: 'Option 1'})
+
+    set_r_select_option('homsOrderDataSelect', 'Option 2')
+
+    expect(r_select_single_value('homsOrderDataSelect')).to eq({label: 'Option 2', value: 'Option 2'})
     expect_widget_presence
   end
 
@@ -112,9 +116,9 @@ feature 'Check select with', js: true do
 
     click_and_wait 'ORD-9'
 
-    expect(page).to have_content  'ORD-9'
-    expect(page).to have_selector "[name='homsOrderNotInVBPVariables']"
-    expect(page).to have_content  'Field with name homsOrderNotInVBPVariables not defined in BP variables'
+    expect(page).to have_content 'ORD-9'
+    expect_r_select_presence('homsOrderNotInVBPVariables')
+    expect(page).to have_content 'Field with name homsOrderNotInVBPVariables not defined in BP variables'
     expect_widget_presence
   end
 
@@ -126,8 +130,8 @@ feature 'Check select with', js: true do
     click_and_wait ('ORD-12')
 
     expect(page).to have_content 'ORD-12'
-    expect(page).to have_selector("[name='homsOrderDataSelect']")
-    expect(select2_text('homsOrderDataSelect')).to eq first_value
+    expect_r_select_presence('homsOrderDataSelect')
+    expect(r_select_single_value('homsOrderDataSelect')[:label]).to eq first_value
     expect(page).to have_no_content 'Field with name homsOrderNotInVBPVariables not defined in BP variables'
     expect_widget_presence
   end