+++ CHANGELOG.md
@@ -17,6 +17,7 @@ v1.5.0 [unreleased]
 -   [#184](https://github.com/latera/homs/pull/184) Add `base_url` option to main config provides custom prefix for all paths in app. Fixes [#148](https://github.com/latera/homs/issues/148).    
 -   [#185](https://github.com/latera/homs/pull/185) Speed up docker build.
 -   [#188](https://github.com/latera/homs/pull/188) Use `.env` file in docker-compose.
+-   [#183](https://github.com/latera/homs/pull/183) Replace `Select2` with `react-select` in widget.
 
 ### Bugfixes
 -   [#170](https://github.com/latera/homs/pull/170) Use active support logger as base.